--- a/src/emscripten-optimizer/CMakeLists.txt
+++ b/src/emscripten-optimizer/CMakeLists.txt
@@ -1,0 +1,6 @@
+SET(optimizer_SOURCES
+  parser.cpp
+  simple_ast.cpp
+  optimizer-shared.cpp
+)
+ADD_LIBRARY(optimizer STATIC ${optimizer_SOURCES})