--- a/src/pass.cpp
+++ b/src/pass.cpp
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include <pass.h>
+#include "pass.h"
 
 namespace wasm {