--- a/src/support/CMakeLists.txt
+++ b/src/support/CMakeLists.txt
@@ -1,0 +1,8 @@
+SET(support_SOURCES
+  bits.cpp
+  colors.cpp
+  command-line.cpp
+  file.cpp
+  safe_integer.cpp
+)
+ADD_LIBRARY(support STATIC ${support_SOURCES})