@kripken 2016-03-31
Looks like this fails on the bots. It can't find a pass, which implies that maybe you aren't linking in all the files under `src/passes`/ into the binaryen shell, or another tool?