@shibukawa 2016-04-01
I read passes code and understood. Linking code has side-effects because source codes in pases folder register theirselves. So my initial approach was wrong (it links all codes as a static library).

Now libpass.a contains minimum codes. The result becomes 50s -> 43s.