@kripken 2016-04-01
Thanks, looks good to me know. But might be good if someone that knows cmake better than me (which is pretty much everyone :) ) could take a look too.

Also, please join the [community group](https://www.w3.org/community/webassembly/participants) before we merge, if you haven't already.