@jfbastien 2016-04-01
I'm OK if https://github.com/WebAssembly/binaryen/pull/302#discussion_r58278110 is done separately or here. I don't want to block this change on it, but I think it'll keep things sane.