@kripken 2016-04-12
How about if we have a single `binaryen.so` shared library, that all the tools would be linked with dynamically? That would be good for other projects using binaryen as well, they would just link with it similarly.