@jfbastien 2016-04-12
> How about if we have a single binaryen.so shared library, that all the tools would be linked with dynamically? That would be good for other projects using binaryen as well, they would just link with it similarly.

I think it's better to do small static libraries where possible: it keeps code sane, prevents introduction of odd dependencies, and keeps fast build time. Not that dylibs are bad either, it just seems to be a slightly unrelated thing: the end tools build faster, but you still need to rebuild the `.so`.