@jfbastien 2016-04-12
> Small static libs internally sounds ok, but are you fine with also having a single `.so` for external projects? e.g. the program in examples/, I'd like it to just link to `libbinaryen` and not to multiple static libs.

Yeah that sounds good, I'd tackle it after the static libs though.