@shibukawa 2016-04-14
I updated my PR to use ADD_SUBDIRECTORY.