@jfbastien 2016-04-14
Awesome, I think that's a pretty good improvement on its own. We can probably do more in that vein afterwards.