@kripken 2016-04-18
Looks nice, thanks. Please just rebase so this can be merged, and also I think you need to join the [WebAssembly w3c group](https://www.w3.org/community/webassembly/participants).