@dschuff 2016-04-21
@shibukawa Do you plan on rebasing this? If not, do you mind if I do that and land it? I have another refactoring that would be much nicer if this were in with it.