@shibukawa 2016-04-23
If you are making better solution, feel free to discard my commit.