@kripken 2016-05-09
Ok, another solution has landed for this, so I think this can be closed.